scripts/requirements.txt
sentence-transformers==2.2.2
pinecone-client==3.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.6.1
